Avoid use of std::env::set_var
Starting in Rust edition 2024 this function is categorized as unsafe to
call in a multithreaded program. Because tests run in multiple threads
we can't safely modify the environment variables. In fact if we had
multiple tests that used environment variables they could overwrite each
other, causing intermittent test failures.

Therefore the safest option is to avoid modifying the evironment at all
in the tests, and passing the list of environment variables into
parse_config as an argument, thereby "mocking" the environment.