Move Box inside DnsErrorKind
I see from the git history that the reason we are boxing this error is
that ResolveError may be large. However then we only need to box the
DnsResolve variant, this saves a level of indirection in the
UnexpectedResponse case without increasing the enum size.

I think it's questionable whether we need a Box at all, since without
boxing the enum size is only 32 bytes.